Split out connect function
